* move classes into own files * move DialogEvents into top level enum instead of embeeded in dialogContext
