+++ libraries/Microsoft.Bot.Builder.Dialogs.Adaptive/AdaptiveEvents.cs
@@ -1,0 +1,11 @@
+﻿// Licensed under the MIT License.
+// Copyright (c) Microsoft Corporation. All rights reserved.
+
+namespace Microsoft.Bot.Builder.Dialogs.Adaptive
+{
+    public class AdaptiveEvents : DialogEvents
+    {
+        public const string RecognizedIntent = "recognizedIntent";
+        public const string UnknownIntent = "unknownIntent";
+    }
+}
